--- src/main/java/enums/NegativeTokens.java
@@ -1,22 +1,0 @@
-package enums;
-
-public enum NegativeTokens {
-    WHITESPACE(' '),
-    EXCLAMATIONMARK('!'),
-    DOUBLEQUOTE('"'),
-    NUMBERSIGN('#'),
-    DOLLAR('$'),
-    PERCENT('%'),
-    AMPERSAND('&'),
-    DOUBLEBACKSLASH('\\'),
-    LEFTPARENTHESIS('('),
-    RIGHTPARENTHESIS(')');
-
-
-    public final char value;
-
-    NegativeTokens(char negativeTokens) {
-        this.value = negativeTokens;
-    }
-}
-
