--- src/main/java/enums/Salutation.java
@@ -1,13 +1,0 @@
-package enums;
-
-public enum Salutation {
-    MVR("Mvr"),
-    DHR("Dhr");
-
-    public final String value;
-
-    Salutation(String salutation) {
-        this.value = salutation;
-    }
-}
-
