--- src/main/java/enums/Vat.java
@@ -1,13 +1,0 @@
-package enums;
-
-public enum Vat {
-    HIGH("Hoog"),
-    LOW("Laag"),
-    NONE("0");
-
-    public final String value;
-
-    Vat(String vat) {
-        this.value = vat;
-    }
-}
