+++ src/main/java/logic/enums/NegativeTokens.java
@@ -1,0 +1,22 @@
+package logic.enums;
+
+public enum NegativeTokens {
+    WHITESPACE(' '),
+    EXCLAMATIONMARK('!'),
+    DOUBLEQUOTE('"'),
+    NUMBERSIGN('#'),
+    DOLLAR('$'),
+    PERCENT('%'),
+    AMPERSAND('&'),
+    DOUBLEBACKSLASH('\\'),
+    LEFTPARENTHESIS('('),
+    RIGHTPARENTHESIS(')');
+
+
+    public final char value;
+
+    NegativeTokens(char negativeTokens) {
+        this.value = negativeTokens;
+    }
+}
+
