+++ src/main/java/logic/enums/Salutation.java
@@ -1,0 +1,13 @@
+package logic.enums;
+
+public enum Salutation {
+    MVR("Mvr"),
+    DHR("Dhr");
+
+    public final String value;
+
+    Salutation(String salutation) {
+        this.value = salutation;
+    }
+}
+
