+++ src/main/java/persistence/people/dao/BaseDao.java
@@ -1,0 +1,14 @@
+package persistence.people.dao;
+
+import persistence.people.connector.PostgresConnector;
+
+import java.io.IOException;
+import java.sql.Connection;
+
+public class BaseDao {
+    private PostgresConnector connector = PostgresConnector.getInstance();
+
+    public Connection getConnection() throws IOException {
+        return connector.getConnection();
+    }
+}
