--- src/main/java/sql/models/Bank.java
@@ -1,19 +1,0 @@
-package sql.models;
-
-public class Bank {
-    private String iban;
-    private String bic;
-
-    public Bank(String iban, String bic) {
-        this.iban = iban;
-        this.bic = bic;
-    }
-
-    public String getIban() {
-        return iban;
-    }
-
-    public String getBic() {
-        return bic;
-    }
-}
