--- src/main/java/sql/models/Name.java
@@ -1,33 +1,0 @@
-package sql.models;
-
-import enums.Salutation;
-
-public class Name {
-    private Salutation salutation;
-    private String firstName;
-    private String lastName;
-    private String middleName;
-
-    public Name(Salutation salutation, String firstName, String lastName, String middleName) {
-        this.salutation = salutation;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.middleName = middleName;
-    }
-
-    public Salutation getSalutation() {
-        return salutation;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-}
